Renamed distributor for indexer
