+++ cmd/observer/node_builder/observer_builder.go
@@ -1474,7 +1474,7 @@ func (builder *ObserverServiceBuilder) BuildExecutionSyncComponents() *ObserverS
 					return nil, fmt.Errorf("failed to create execution data pruner: %w", err)
 				}
 
-				builder.ExecutionIndexer.AddOnBlockHeaderReceivedConsumer(func(header *flow.Header) {
+				builder.ExecutionIndexer.AddOnBlockIndexedConsumer(func(header *flow.Header) {
 					err := builder.ExecutionDataTracker.SetFulfilledHeight(header.Height)
 					if err != nil {
 						node.Logger.Fatal().Err(err).Msg("failed to set fulfilled height")
